+++ src/function/index.test.ts
@@ -1,4 +1,4 @@
-import {expectToMatchColor} from '../../test/unit/test_utils';
+import {expectToMatchColor} from '../../test/lib/util';
 import {createFunction} from './index';
 import Color from '../util/color';
 import Formatted from '../expression/types/formatted';