+++ src/validate/validate.ts
@@ -17,6 +17,7 @@ import validateFilter from './validate_filter';
 import validateLayer from './validate_layer';
 import validateSource from './validate_source';
 import validateLight from './validate_light';
+import validateSky from './validate_sky';
 import validateTerrain from './validate_terrain';
 import validateString from './validate_string';
 import validateFormatted from './validate_formatted';
@@ -24,6 +25,7 @@ import validateImage from './validate_image';
 import validatePadding from './validate_padding';
 import validateVariableAnchorOffsetCollection from './validate_variable_anchor_offset_collection';
 import validateSprite from './validate_sprite';
+import ValidationError from '../error/validation_error';
 
 const VALIDATORS = {
     '*'() {
@@ -41,6 +43,7 @@ const VALIDATORS = {
     'object': validateObject,
     'source': validateSource,
     'light': validateLight,
+    'sky': validateSky,
     'terrain': validateTerrain,
     'string': validateString,
     'formatted': validateFormatted,
@@ -50,17 +53,30 @@ const VALIDATORS = {
     'sprite': validateSprite,
 };
 
-// Main recursive validation function. Tracks:
-//
-// - key: string representing location of validation in style tree. Used only
-//   for more informative error reporting.
-// - value: current value from style being evaluated. May be anything from a
-//   high level object that needs to be descended into deeper or a simple
-//   scalar value.
-// - valueSpec: current spec being evaluated. Tracks value.
-// - styleSpec: current full spec being evaluated.
-
-export default function validate(options) {
+/**
+ * Main recursive validation function used internally.
+ * You should use `validateStyleMin` in the browser or `validateStyle` in node env.
+ * @param options - the options object
+ * @param options.key - string representing location of validation in style tree. Used only
+ * for more informative error reporting.
+ * @param options.value - current value from style being evaluated. May be anything from a
+ * high level object that needs to be descended into deeper or a simple
+ * scalar value.
+ * @param options.valueSpec - current spec being evaluated. Tracks value.
+ * @param options.styleSpec - current full spec being evaluated.
+ * @param options.validateSpec - the validate function itself
+ * @param options.style - the style object
+ * @param options.objectElementValidators - optional object of functions that will be called
+ * @returns an array of errors, or an empty array if no errors are found.
+ */
+export default function validate(options: {
+    key: any;
+    value: any;
+    valueSpec: any;
+    styleSpec: any;
+    validateSpec?: any;
+    style: any;
+    objectElementValidators?: any;}): ValidationError[] {
     const value = options.value;
     const valueSpec = options.valueSpec;
     const styleSpec = options.styleSpec;