+++ src/validate/validate_raster_dem_source.test.ts
@@ -1,5 +1,5 @@
 import validateSpec from './validate';
-import v8 from '../reference/v8.json' assert {type: 'json'};
+import v8 from '../reference/v8.json' with {type: 'json'};
 import validateRasterDEMSource from './validate_raster_dem_source';
 import {RasterDEMSourceSpecification} from '../types.g';