+++ src/visit.ts
@@ -1,5 +1,5 @@
-import Reference from './reference/v8.json' assert {type: 'json'};
-import type {StylePropertySpecification} from './style-spec';
+import Reference from './reference/v8.json' with {type: 'json'};
+import type {StylePropertySpecification} from '.';
 import type {
     StyleSpecification,
     SourceSpecification,